Changed logic of plan subscription route so that user can select multiple products
